Added consistency object route tests
